Moved tests out of mod files
